fixed eslint problems
